+++ univocity-trader-marketsimulator/src/main/java/com/univocity/trader/marketsimulator/MarketSimulatorMain.java
@@ -7,15 +7,7 @@ import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 
-import com.univocity.trader.account.SimpleTradingFees;
-import com.univocity.trader.config.UnivocityConfiguration;
 import com.univocity.trader.config.impl.ConfigFileUnivocityConfigurationImpl;
-import com.univocity.trader.currency.Currencies;
-import com.univocity.trader.guice.UnivocityFactory;
-import com.univocity.trader.notification.OrderExecutionToLog;
-import com.univocity.trader.notification.SimpleStrategyStatistics;
-import com.univocity.trader.simulation.MarketSimulator;
-import com.univocity.trader.strategy.StrategyFactory;
 
 /**
  * @author uniVocity Software Pty Ltd - <a href="mailto:dev@univocity.com">dev@univocity.com</a>
@@ -56,35 +48,9 @@ public class MarketSimulatorMain {
           */
          final String currenciesList = cmd.getOptionValue(CURRENCIES_OPTION);
          if ((null != configFileName) && (null != currenciesList)) {
-            final String[] currencies = Currencies.getInstance().fromList(currenciesList);
             ConfigFileUnivocityConfigurationImpl.setConfigfileName(configFileName);
-            final UnivocityConfiguration univocityConfiguration = UnivocityFactory.getInstance().getUnivocityConfiguration();
-            System.out.println("Strategy: " + univocityConfiguration.getStrategyClass().getName());
-            final MarketSimulator simulation = new MarketSimulator(univocityConfiguration.getSimulationReferenceCurrency());
-            simulation.tradeWith(currencies);
-            simulation.strategies().add(StrategyFactory.getInstance().getStrategySupplier(univocityConfiguration.getStrategyClass()));
-            for (Class<?> clazz : univocityConfiguration.getStrategyMonitorClasses()) {
-               System.out.println("Monitor: " + clazz.getName());
-               simulation.monitors().add(StrategyFactory.getInstance().getStrategyMonitorSupplier(clazz));
-            }
-            simulation.setTradingFees(SimpleTradingFees.percentage(0.1));
-            // simulation.symbolInformation("ADAUSDT").minimumAssetsPerOrder(100.0).priceDecimalPlaces(8).quantityDecimalPlaces(2);
-            // simulation.symbolInformation("BTCUSDT").minimumAssetsPerOrder(0.001).priceDecimalPlaces(8).quantityDecimalPlaces(8);
-            // simulation.symbolInformation("LTCUSDT").minimumAssetsPerOrder(0.1).priceDecimalPlaces(8).quantityDecimalPlaces(8);
-            // simulation.symbolInformation("XRPUSDT").minimumAssetsPerOrder(50.0).priceDecimalPlaces(8).quantityDecimalPlaces(2);
-            // simulation.symbolInformation("ETHUSDT").minimumAssetsPerOrder(0.01).priceDecimalPlaces(8).quantityDecimalPlaces(8);
-            simulation.account().setAmount(univocityConfiguration.getSimulationReferenceCurrency(), 1000.0).minimumInvestmentAmountPerTrade(10.0);
-            // .maximumInvestmentPercentagePerAsset(30.0, "ADA", "ETH")
-            // .maximumInvestmentPercentagePerAsset(50.0, "BTC", "LTC")
-            // .maximumInvestmentAmountPerAsset(200, "XRP")
-            ;
-            simulation.setSimulationStart(univocityConfiguration.getSimulationStart());
-            simulation.setSimulationEnd(univocityConfiguration.getSimulationEnd());
-            simulation.listeners().add(new OrderExecutionToLog());
-            final SimpleStrategyStatistics stats = new SimpleStrategyStatistics();
-            simulation.listeners().add(stats);
-            simulation.run();
-            stats.printTradeStats();
+            final MarketSimulatorRunner marketSimulatorRunner = new MarketSimulatorRunner();
+            marketSimulatorRunner.simulate(currenciesList);
          }
       } catch (final Exception e) {
          e.printStackTrace();
